I also got rid of a bunch of assets that were a bit too large.

Conflicts:
	data/assets/Maps/BigBox.tmx
	data/assets/Maps/The_Match.tmx
	src/Entities/AIEntity.java
	src/Entities/Football.java
	src/Graphics/Camera/FreeCamera.java
	src/States/Game/Adventure/AdventureMode.java
	src/States/Game/RaceMode/RaceMode.java
	src/States/Game/StateGame.java
	src/World/sWorld.java

+++ src/Entities/SpatBlock.java
@@ -59,7 +59,8 @@ public class SpatBlock extends Entity {
                 {
                     
                 }
-                kill(CauseOfDeath.eMundane, null);
+                if (mBody.getFixtureList().getFilterData().categoryBits != (1 << BodyCategories.eGum.ordinal()))
+                    kill(CauseOfDeath.eMundane, null);
                 break;
             }
         }
